+++ app/controllers/mappings_controller.rb
@@ -1,9 +1,9 @@
 class MappingsController < ApplicationController
   before_action :authenticate_user!
-  before_action :set_source_with_edit_data_source_mappings, only: [ :automap_popup, :create ]
-  before_action :redirect_without_source, only: [ :automap_popup, :create ]
-  before_action :set_mapping, only: [ :expanded, :show, :edit, :destroy ]
-  before_action :redirect_without_mapping, only: [ :expanded, :show, :edit, :destroy ]
+  before_action :set_source_with_edit_data_source_mappings, only: [ :automap_popup, :show, :edit, :create, :update, :destroy ]
+  before_action :redirect_without_source, only: [ :automap_popup, :show, :edit, :create, :update, :destroy ]
+  before_action :set_mapping, only: [ :expanded, :show, :edit, :update, :destroy ]
+  before_action :redirect_without_mapping, only: [ :expanded, :show, :edit, :update, :destroy ]
 
   def automap_popup
   end
@@ -19,7 +19,7 @@ class MappingsController < ApplicationController
     if @mapping
       chart_params = {}
       if @mapping.concept.continuous? or @mapping.concept.date?
-        chart_params = { title: @mapping.concept.human_name, width: 381, height: 300, units: @mapping.human_units, legend: 'none' }
+        chart_params = { title: @mapping.concept.human_name, width: 381, height: 300, units: @mapping.concept.human_units, legend: 'none' }
       elsif @mapping.concept.categorical? or @mapping.concept.boolean?
         chart_params = { title: @mapping.concept.human_name, width: 381, height: 250 }
       end
@@ -39,7 +39,7 @@ class MappingsController < ApplicationController
   def expanded
     chart_params = {}
     if @mapping.concept.continuous? or @mapping.concept.date?
-      chart_params = { title: @mapping.concept.human_name, width: 680, height: 300, units: @mapping.human_units, legend: 'none' }
+      chart_params = { title: @mapping.concept.human_name, width: 680, height: 300, units: @mapping.concept.human_units, legend: 'none' }
     elsif @mapping.concept.categorical? or @mapping.concept.boolean?
       chart_params = { title: @mapping.concept.human_name, width: 450, height: 250 }
     end
@@ -55,7 +55,6 @@ class MappingsController < ApplicationController
 
   # GET /mappings/1
   def show
-    @mapping.set_status!(current_user)
   end
 
   # GET /mappings/1/edit
@@ -66,86 +65,35 @@ class MappingsController < ApplicationController
   def create
     @concept = Concept.find_by_id(params[:concept_id])
     if @concept
-      @mapping = @source.mappings.where( table: params[:table], column: params[:column] ).first_or_create
-      @mapping.update(units: @concept.units, concept_id: @concept.id, deleted: false)
-      @mapping.reload
-      @mapping.set_status!(current_user)
+      @mapping = @source.mappings.create( table: params[:table], column: params[:column], concept_id: @concept.id )
+      @mapping.automap(current_user)
     end
   end
 
-  # TODO: Rewrite to use just def update (refactor)
-  def update_multiple
-    @source = current_user.all_sources.find_by_id(params[:source_id])
-    source = Source.find_by_id(params[:source_id])
-    @source = source if (not @source) and source and source.user_has_action?(current_user, "edit data source mappings")
+  # PATCH /mappings/1
+  def update
+    (mapping_params[:column_values] || []).each do |column_value_hash|
+      column_value = column_value_hash[:is_null] == 'true' ? nil : column_value_hash[:column_value]
+      value = column_value_hash[:value]
+      val_mapping = @source.mappings.where( table: @mapping.table, column: @mapping.column, column_value: (column_value == nil ? 'NULL' : column_value.to_s) ).first_or_create
 
-    params[:mapping_ids] = [params[:selected_mapping_id].to_i] unless params[:selected_mapping_id].blank?
-
-    if not @source or params[:mapping_ids].blank?
-      render nothing: true
-    else
-      @mappings = @source.mappings.find(params[:mapping_ids])
-      @mappings.each do |mapping|
-        @mapping = mapping
-        current_time = Time.now
-
-        if params[:mappings][@mapping.id.to_s]
-
-          if params[:mappings][@mapping.id.to_s][:mapping_column_values]
-            params[:mappings][@mapping.id.to_s][:mapping_column_values].each do |column_value, value_hash|
-              value = value_hash[:value]
-              if value_hash[:is_null] == 'true'
-                column_value = nil
-              elsif column_value == "value_&lt;&#33;&#91;CDATA&#91;&#93;&#93;&gt;"
-                column_value = ''
-              else
-                column_value = column_value[6..-1]  # Strip off "value_" from front of values
-              end
-              # .gsub('&lt;', '<').gsub('&gt;', '>')
-              column_value = column_value.gsub('&#33;', '!').gsub('&#91;', '[').gsub('&#93;', ']').gsub('&#32;', ' ') if column_value
-              val_mapping = @source.mappings.where( table: @mapping.table, column: @mapping.column, column_value: (column_value == nil ? 'NULL' : column_value.to_s) ).first_or_create
-
-              value = value.blank? ? nil : value
-              if @mapping.concept
-                if Concept.find_by_id(value)
-                  val_mapping.update_attributes(concept_id: value, value: nil, status: 'mapped', deleted: false)
-                else
-                  val_mapping.update_attributes(concept_id: @mapping.concept_id, value: value, status: (value == nil) ? 'unmapped' : 'mapped', deleted: false)
-                end
-              end
-            end
-          end
-
-          @mapping.reload
-          # TODO: Remove Mappings that no longer exist in the underlying data source
-          # @mapping.database_concept_column_values.where(["time_stamp != ?", current_time]).each {|dccv| dccv.destroy}
-
-          params[:mappings][@mapping.id.to_s].reject!{|key, value| ['mapping_column_values'].include?(key.to_s)}
-
-          if @mapping.update_attributes(params[:mappings][@mapping.id.to_s])
-            @mapping.set_status!(current_user)
-            flash[:notice] = 'Mapping was successfully updated.'
-          end
+      value = value.blank? ? nil : value
+      if @mapping.concept
+        if Concept.find_by_id(value)
+          val_mapping.update( concept_id: value, value: nil )
+        else
+          val_mapping.update( concept_id: @mapping.concept_id, value: value )
         end
       end
-
-      table_div_name = params[:table].gsub(/[^\w\-]/, '_')
-      result_hash = @source.table_columns(current_user, params[:table])
-      @columns = result_hash[:result]
-      all_column_names = @columns.collect{|a| a[:column]}
     end
+    render 'show'
   end
 
   # DELETE /mappings/1.js
   def destroy
     @column = @mapping.column
-    @source = @mapping.source
-    @table = @mapping.table
-    @mapping_id = @mapping.id
-
-    @mapping.destroy
-
-    # TODO: Update number of mapped concepts
+    params[:table] = @mapping.table
+    @source.mappings.where( table: @mapping.table, column: @mapping.column ).destroy_all
     render 'new'
   end
 
@@ -160,4 +108,10 @@ class MappingsController < ApplicationController
       empty_response_or_root_path unless @mapping
     end
 
+    def mapping_params
+      params[:mapping] ||= { blank: true }
+      params.require(:mapping).permit(
+        { :column_values => [ :value, :column_value, :is_null ] }
+      )
+    end
 end