+++ db/migrate/20130627153708_remove_deleted_and_status_and_units_from_mappings.rb
@@ -1,0 +1,14 @@
+class RemoveDeletedAndStatusAndUnitsFromMappings < ActiveRecord::Migration
+  def up
+    Mapping.where( deleted: true ).delete_all
+    remove_column :mappings, :deleted
+    remove_column :mappings, :status
+    remove_column :mappings, :units
+  end
+
+  def down
+    add_column :mappings, :deleted, :boolean, null: false, default: false
+    add_column :mappings, :status, :string
+    add_column :mappings, :units, :string
+  end
+end