+++ test/models/source_test.rb
@@ -5,11 +5,11 @@ class SourceTest < ActiveSupport::TestCase
     assert_equal true, sources(:two).user_has_action_group?(users(:admin), "All Write")
     assert_equal false, sources(:two).user_has_action_group?(users(:valid), "All Write")
   end
-  
+
   test "should show false if the action group does not exist" do
     assert_equal false, sources(:two).user_has_action_group?(users(:admin), "All Code")
   end
-  
+
   test "should show join conditions given an array of tables" do
     result_hash = sources(:two).join_conditions(['table', 'table2', 'table3'], users(:admin))
     assert result_hash.kind_of?(Hash)
@@ -18,10 +18,5 @@ class SourceTest < ActiveSupport::TestCase
     assert result_hash[:result].size > 0
     assert result_hash[:errors].size > 0
   end
-  
-  test "should generate derived mappings" do
-    sources(:two).generate_derived_mappings!
-    assert_equal 1, sources(:two).mappings.status('derived').size
-  end
-  
+
 end